// src/model/connection.ts
export interface IConnection {
    // readonly host: string;
    // readonly user: string;
    // readonly password?: string;
    // readonly port: string;
    // readonly database?: string;
    // multipleStatements?: boolean;
    // readonly certPath: string;
    readonly name: string;
    readonly connectionString: string;
}